+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 -   Internal: Use [`@backtrack/preset-node`](https://github.com/chrisblossom/backtrack-preset-node)
+-   Update [`resolve-with-prefix`](https://github.com/chrisblossom/resolve-with-prefix)
 
 ## [1.0.12] - 2019-01-02